Atualizando arquivos

+++ src/main/java/com/example/CadastroDeNinjas/Missoes/MissoesModel.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Entity
 @Table(name = "tb_missoes")
 @Data
@@ -18,7 +19,7 @@ public class MissoesModel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String nome;
